Switch to workspace directory
