+++ rust/account/Cargo.toml
@@ -16,10 +16,10 @@ name = "account"
 path = "benches/account.rs"
 harness = false
 
-[dependencies.snarkvm-wasm]
-git = "https://github.com/AleoHQ/snarkVM.git"
-version = "0.9.3"
-features = ["parallel"]
+[dependencies.snarkvm]
+workspace = true
+default-features = false
+features = ["parallel", "wasm"]
 
 [dev-dependencies.rand]
 version = "0.8.0"
