+++ rust/account/src/lib.rs
@@ -16,8 +16,27 @@
 
 #![forbid(unsafe_code)]
 
-pub mod account;
-pub use account::*;
-
 #[cfg(test)]
 pub mod tests;
+
+use snarkvm::wasm::{
+    account::{
+        Address as AleoAddress,
+        PrivateKey as AleoPrivateKey,
+        Signature as AleoSignature,
+        ViewKey as AleoViewKey,
+    },
+    network::Testnet3,
+    program::{Ciphertext as AleoCiphertext, Record as AleoRecord},
+};
+
+pub use snarkvm::wasm::{network::Environment, FromBytes, PrimeField, ToBytes};
+
+pub type CurrentNetwork = Testnet3;
+
+pub type Address = AleoAddress<CurrentNetwork>;
+pub type PrivateKey = AleoPrivateKey<CurrentNetwork>;
+pub type Signature = AleoSignature<CurrentNetwork>;
+pub type ViewKey = AleoViewKey<CurrentNetwork>;
+
+pub type Record = AleoRecord<CurrentNetwork, AleoCiphertext<CurrentNetwork>>;
